allow a facing object to be compared
this allows comparisons with other facing objects and symbols. you can't
however compare a symbol to a facing object (symbol == facing) as that
would require monkey patching Symbol, which is yuck.

it's a shame that the 'coerce' method of extending the number types
doesn't exist for symbols (at least in so far as i'm aware).